metrics: Rename monitor package to metrics
Makes more sense by the abstractions I intend
to expose from it.
